Refactored federation-utils and moved entry.graphql to common.graphql for code cleaning
