--- GG_generator/src/Back-End/Service/src/graphql/Service.entry.graphql.hbs
@@ -1,4 +1,0 @@
-"""
-A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.
-"""
-scalar DateTime
