+++ GG_utilities/src/graphql/QueryArgs.ts
