+++ GG_utilities/src/index.ts
@@ -1,11 +1,21 @@
-import resolvers from './resolvers';
+import { loadFilesSync } from '@graphql-tools/load-files';
+import path from 'path';
+
+// PACKAGE IMPORTS
+import resolvers from './graphql/resolvers';
 
 import { cryptObject, decryptObject } from './security/Hashing';
 import { verifyToken, signToken } from './security/JWT';
 
-import { PaginationAndSearchArgs } from './types/QueryArgs';
+import { PaginationAndSearchArgs } from './graphql/QueryArgs';
 
 import authDirective from './directives/auth';
+
+const typeDefs = loadFilesSync(path.join(__dirname, '.'), {
+  recursive: true,
+  extensions: ['graphql'],
+  ignoreIndex: true,
+});
 
 // EXPORTS
 export {
@@ -16,4 +26,5 @@ export {
   verifyToken,
   signToken,
   authDirective,
+  typeDefs as commonTypeDefs,
 };
