+++ server/services/federation-auth/package.json
@@ -71,7 +71,7 @@
     "dotenv-safe": "^8.2.0",
     "express": "^4.17.1",
     "faker": "^5.5.3",
-    "federation-utils": "../../bin/federation-utils/federation-utils-v1.0.19.tgz",
+    "federation-utils": "../../bin/federation-utils/federation-utils-v1.0.23.tgz",
     "graphql": "^15.5.0",
     "graphql-middleware": "^6.1.10",
     "graphql-scalars": "^1.12.0",
